+++ packages/hydrated_cubit/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.3
+
+- **BREAKING**:
+  - Rename `HydratedStorage` to `Storage`
+  - Rename `HydratedCubitStorage` to `HydratedStorage`
+  - Rename `getInstance` to `build`
+
 # 0.0.2
 
 - docs: minor improvements to README